LPS-109473 Extract class
